repsoitory 중복 오류 해결
